ingest/requirements.txt
boto3>=1.4.7
boto>=2.48.0
pyyaml>=3.10
argparse>=1.1
